backend/requirements.txt
pandas==1.1.0
requests==2.24.0
beautifulsoup4==4.6.3
lxml==4.5.1
yarl==1.4.2
asyncio==3.4.3
aiohttp==3.6.2
pytest==5.4.3
google-cloud-bigquery==1.25.0
google-cloud==0.34.0
Flask==1.1.2
google-cloud-logging==1.15.0
pandas-gbq==0.13.2
pyarrow==1.0.0
google-cloud-storage==1.25.0 
firebase-admin==4.3.0
google-cloud-firestore==1.8.1
grpcio==1.31.0